Rename DataLoader -> DataCreator, DataGenerator -> DataLoader
